changed levels

+++ src/com/ggwp/snake/Obstructions.java
@@ -2,17 +2,17 @@ package com.ggwp.snake;
 
 public class Obstructions {
 	
-	int top;
-	int bottom;
-	int left;
-	int right;
-	int radius;
-	int circle_x;
-	int circle_y;
+	float top;
+	float bottom;
+	float left;
+	float right;
+	float radius;
+	float circle_x;
+	float circle_y;
 	String shape;
 	
 	
-	public Obstructions(int top, int bottom, int left, int right) {
+	public Obstructions(float top, float bottom, float left, float right) {
 			shape = "Rectangle";
 			this.top = top;
 			this.bottom = bottom;
@@ -20,62 +20,62 @@ public class Obstructions {
 			this.right = right;
 	}
 
-	public Obstructions(int radius,int x,int y) {
+	public Obstructions(float radius,float x,float y) {
 		shape = "Circle";
 		this.circle_x = x;
 		this.circle_y = y;
 		this.radius = radius;
 	}
 	
-	public int getRadius() {
+	public float getRadius() {
 		return radius;
 	}
 
-	public void setRadius(int radius) {
+	public void setRadius(float radius) {
 		this.radius = radius;
 	}
 
-	public int getCircle_x() {
+	public float getCircle_x() {
 		return circle_x;
 	}
 
-	public void setCircle_x(int circle_x) {
+	public void setCircle_x(float circle_x) {
 		this.circle_x = circle_x;
 	}
 
-	public int getCircle_y() {
+	public float getCircle_y() {
 		return circle_y;
 	}
 
-	public void setCircle_y(int circle_y) {
+	public void setCircle_y(float circle_y) {
 		this.circle_y = circle_y;
 	}
 
 	public String getShape() {
 		return shape;
 	}
-	public int getTop() {
+	public float getTop() {
 		return top;
 	}
-	public void setTop(int top) {
+	public void setTop(float top) {
 		this.top = top;
 	}
-	public int getBottom() {
+	public float getBottom() {
 		return bottom;
 	}
-	public void setBottom(int bottom) {
+	public void setBottom(float bottom) {
 		this.bottom = bottom;
 	}
-	public int getLeft() {
+	public float getLeft() {
 		return left;
 	}
-	public void setLeft(int left) {
+	public void setLeft(float left) {
 		this.left = left;
 	}
-	public int getRight() {
+	public float getRight() {
 		return right;
 	}
-	public void setRight(int right) {
+	public void setRight(float right) {
 		this.right = right;
 	}
